[TASK-57] BOM 기준정보 관리 고도화 및 UX/UI 개선

+++ backend/src/main/java/com/ssafy/demo_app/api/production/dto/WorkOrderMaterialRequirementResponse.java
@@ -3,8 +3,6 @@ package com.ssafy.demo_app.api.production.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.math.BigDecimal;
-
 @Getter
 @AllArgsConstructor
 public class WorkOrderMaterialRequirementResponse {
@@ -13,7 +11,7 @@ public class WorkOrderMaterialRequirementResponse {
     private String itemName;
     private String itemType;
     private String unit;
-    private BigDecimal bomQuantity;
+    private Integer bomQuantity;
     private Integer requiredQty;
     private Integer issuedQty;
     private Integer availableQty;
